SCPI accepte les parts fractionnées — Room v8→v9
sharesCount passe de Int à Double (REAL en SQLite) pour permettre
des valeurs comme 2,2 parts. Migration v8→v9 recrée la table
scpi_investments en préservant les données existantes.

+++ app/src/main/java/com/dibitara/app/data/local/entity/ScpiInvestmentEntity.kt
@@ -11,7 +11,7 @@ data class ScpiInvestmentEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0,
     val label: String,
-    val sharesCount: Int,
+    val sharesCount: Double,
     val shareValueCents: Long,
     val monthlyContributionCents: Long,
     val currency: String,
